Audio! Error Handling & No More Warnings

+++ winapi/graphics/buffers/IndexBuffer.h
@@ -11,12 +11,12 @@ class IndexBuffer {
 		Microsoft::WRL::ComPtr<ID3D11Buffer> m_pIndexBuffer;
 		Microsoft::WRL::ComPtr<ID3D11DeviceContext>& m_pDeviceContextRef;
 		
-		uint64_t m_nBytes;
+		size_t m_nBytes;
 
 	public:
 		IndexBuffer(const Microsoft::WRL::ComPtr<ID3D11Device>& pDeviceRef, Microsoft::WRL::ComPtr<ID3D11DeviceContext>& pDeviceContextRef, const std::vector<uint32_t>& indices);
 
-		uint64_t getSize() const;
+		size_t getSize() const;
 
 		void Bind() const noexcept;
 };
